Delete SlidingWindowOracle - This oracle costs way too much to maintain, so support is being dropped - Also replaced the sliding-window script with a cleaner consult-aggregate script
